wip: refactoring proofs generation (curve KO)
