--- x-chain/inputs.json
@@ -1,50 +1,0 @@
-[
-    {
-        "protocol": "curve",
-        "gauge_controller": "0x2F50D538606Fa9EDD2B11E2446BEb18C9D5846bB",
-        "start_block": 10647875,
-        "use_state_sender": true,
-        "platforms": [
-            {
-                "chainId": 42161,
-                "address": "0xB854cF650F5492d23e52cb2A7a58B787fC25B0Bb"
-            }
-        ]
-    },
-    {
-        "protocol": "balancer",
-        "gauge_controller": "0xC128468b7Ce63eA702C1f104D55A2566b13D3ABD",
-        "start_block": 14457014,
-        "use_state_sender": false,
-        "platforms": [
-            {
-                "chainId": 42161,
-                "address": "0xFf276AB161f48f6DBa99dE4601f9a518D1d903f9"
-            }
-        ]
-    },
-    {
-        "protocol": "frax",
-        "gauge_controller": "0x3669C421b77340B2979d1A00a792CC2ee0FcE737",
-        "start_block": 14052749,
-        "use_state_sender": false,
-        "platforms": [
-            {
-                "chainId": 42161,
-                "address": "0x4941c004dC4Ae7bcb74B404fbd4ff07Dc32e3ecc"
-            }
-        ]
-    },
-    {
-        "protocol": "fxn",
-        "gauge_controller": "0xe60eB8098B34eD775ac44B1ddE864e098C6d7f37",
-        "start_block": 18156185,
-        "use_state_sender": false,
-        "platforms": [
-            {
-                "chainId": 42161,
-                "address": "0xE5cE02443942B006d0851d6e73d9dbEeE743b88d"
-            }
-        ]
-    }
-]
